+++ tests/experimental.test.js
@@ -3,7 +3,7 @@ import { run, html, css } from './util/run'
 test('experimental universal selector improvements (box-shadow)', () => {
   let config = {
     experimental: 'all',
-    content: [{ raw: html`<div class="shadow resize"></div>` }],
+    content: [{ raw: html`<div class="resize shadow"></div>` }],
     corePlugins: { preflight: false },
   }
 
@@ -39,7 +39,7 @@ test('experimental universal selector improvements (box-shadow)', () => {
 test('experimental universal selector improvements (pseudo hover)', () => {
   let config = {
     experimental: 'all',
-    content: [{ raw: html`<div class="hover:shadow resize"></div>` }],
+    content: [{ raw: html`<div class="resize hover:shadow"></div>` }],
     corePlugins: { preflight: false },
   }
 
@@ -75,7 +75,7 @@ test('experimental universal selector improvements (pseudo hover)', () => {
 test('experimental universal selector improvements (multiple classes: group)', () => {
   let config = {
     experimental: 'all',
-    content: [{ raw: html`<div class="group-hover:shadow resize"></div>` }],
+    content: [{ raw: html`<div class="resize group-hover:shadow"></div>` }],
     corePlugins: { preflight: false },
   }
 
@@ -111,7 +111,7 @@ test('experimental universal selector improvements (multiple classes: group)', (
 test('experimental universal selector improvements (child selectors: divide-y)', () => {
   let config = {
     experimental: 'all',
-    content: [{ raw: html`<div class="divide-y resize"></div>` }],
+    content: [{ raw: html`<div class="resize divide-y"></div>` }],
     corePlugins: { preflight: false },
   }
 
@@ -138,7 +138,7 @@ test('experimental universal selector improvements (child selectors: divide-y)',
 test('experimental universal selector improvements (hover:divide-y)', () => {
   let config = {
     experimental: 'all',
-    content: [{ raw: html`<div class="hover:divide-y resize"></div>` }],
+    content: [{ raw: html`<div class="resize hover:divide-y"></div>` }],
     corePlugins: { preflight: false },
   }
 
@@ -166,7 +166,7 @@ test('experimental universal selector improvements (#app important)', () => {
   let config = {
     experimental: 'all',
     important: '#app',
-    content: [{ raw: html`<div class="shadow divide-y resize"></div>` }],
+    content: [{ raw: html`<div class="resize divide-y shadow"></div>` }],
     corePlugins: { preflight: false },
   }